requirements.txt
fastapi[standard]       # Pour créer l'API Web
uvicorn[standard]       # Serveur ASGI pour exécuter l'API
python-dotenv           # Gestion des variables d'environnement
jinja2                  # Moteur de templates (si besoin d'un frontend avec FastAPI)
python-multipart        # Pour gérer les fichiers envoyés via API
pandas                  # Pour manipuler les données (dataset)
scikit-learn            # Pour entraîner et utiliser le modèle de classification
pytest                  # Pour les tests unitaires
pytest-asyncio          # Pour tester les endpoints asynchrones
httpx                   # Client HTTP pour les tests d'intégration
sqlite-utils            # Pour interagir facilement avec SQLite (historique des prédictions)
plotly
python-dotenv